# app/api/v1/patch-notes/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { listPatchNotes, createPatchNote } from '@/lib/api/patch-notes';

export async function GET(request: NextRequest) {
  const searchParams = request.nextUrl.searchParams;
  const limit = Math.min(parseInt(searchParams.get('limit') || '50'), 100);
  const offset = parseInt(searchParams.get('offset') || '0');

  const result = await listPatchNotes(limit, offset);

  if (!result.success) {
    return NextResponse.json({ error: result.error }, { status: 500 });
  }

  return NextResponse.json(result.data);
}

export async function POST(request: NextRequest) {
  const body = await request.json();

  const result = await createPatchNote(body);

  if (!result.success) {
    return NextResponse.json({ error: result.error }, { status: 500 });
  }

  return NextResponse.json(result.data, { status: 201 });
}